implement optional spoiler protection
Add a configuration option spoiler_protection_days, which allows the admin
to set a number of days (looking backward) where the spoiler function is
disabled. Setting this to 1 protects the current day, setting this to 2
protects the current and the previous day, etc. Setting it to 0 disables
protection (allowing all puzzles to be spoiled) and any negative value
deactivates spoiler functionality completely.

By default, spoiler protection is disabled (equivalent to 0).